+++ CHANGELOG.md
@@ -10,10 +10,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
    Unreleased section, uncommenting the header as necessary.
 -->
 
-<!-- ## Unreleased -->
+## Unreleased
 
 <!-- ### Changed -->
-<!-- ### Added -->
+### Added
+* Added the _scheduleUpdate method to allow control of update timing.
 <!-- ### Removed -->
 <!-- ### Fixed -->